recorder-bot/requirements.txt
asn1crypto==0.24.0
configparser==3.5.0b2
cryptography==2.6.1
entrypoints==0.3
enum34==1.1.6
ipaddress==1.0.17
keyring==17.1.1
keyrings.alt==3.1.1
pycrypto==2.6.1
PyGObject==3.30.4
pyxdg==0.25
SecretStorage==2.3.1
six==1.12.0
virtualenv==15.1.0
aiohttp==3.7.4
matrix-nio[e2e]==0.18.6
markdown==3.3.4
PyYAML==5.4.1
tinydb==4.5.1
requests==2.26.0
transformers==4.10.2
torch==1.9.0
